+++ python/GafferSceneUI/LightEditor.py
@@ -154,7 +154,7 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 		GafferSceneUI.LightEditor.registerColumn(
 			rendererKey,
 			".".join( x for x in [ parameter.shader, parameter.name ] if x ),
-			lambda scene, editScope : _GafferSceneUI._LightEditorInspectorColumn(
+			lambda scene, editScope : GafferSceneUI.Private.InspectorColumn(
 				GafferSceneUI.Private.ParameterInspector( scene, editScope, rendererKey, parameter ),
 				columnName if columnName is not None else ""
 			),
@@ -175,7 +175,7 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 		GafferSceneUI.LightEditor.registerColumn(
 			rendererKey,
 			".".join( x for x in [ parameter.shader, parameter.name ] if x ),
-			lambda scene, editScope : _GafferSceneUI._LightEditorInspectorColumn(
+			lambda scene, editScope : GafferSceneUI.Private.InspectorColumn(
 				GafferSceneUI.Private.ParameterInspector( scene, editScope, shaderAttribute, parameter ),
 				columnName if columnName is not None else ""
 			),
@@ -189,7 +189,7 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 		GafferSceneUI.LightEditor.registerColumn(
 			rendererKey,
 			attributeName,
-			lambda scene, editScope : _GafferSceneUI._LightEditorInspectorColumn(
+			lambda scene, editScope : GafferSceneUI.Private.InspectorColumn(
 				GafferSceneUI.Private.AttributeInspector( scene, editScope, attributeName ),
 				displayName
 			),
@@ -199,7 +199,7 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 	# Registers a column in the Light Editor.
 	# `inspectorFunction` is a callable object of the form
 	# `inspectorFunction( scene, editScope )` returning a
-	# `GafferSceneUI._LightEditorInspectorColumn` object.
+	# `GafferSceneUI.Private.InspectorColumn` object.
 	@classmethod
 	def registerColumn( cls, rendererKey, columnKey, inspectorFunction, section = None ) :
 
@@ -333,19 +333,17 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 		inspectors = {}
 		inspections = []
 
-		with Gaffer.Context( self.context() ) as context :
-
-			for selection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
-				if not isinstance( column, _GafferSceneUI._LightEditorInspectorColumn ) :
-					continue
-				for pathString in selection.paths() :
-					path = GafferScene.ScenePlug.stringToPath( pathString )
-
-					context["scene:path"] = path
+		path = pathListing.getPath().copy()
+		for selection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
+			if not isinstance( column, GafferSceneUI.Private.InspectorColumn ) :
+				continue
+			for pathString in selection.paths() :
+				path.setFromString( pathString )
+				with path.inspectionContext() :
 					inspection = column.inspector().inspect()
 
 					if inspection is not None :
-						inspectors.setdefault( column.inspector(), {} )[path] = inspection
+						inspectors.setdefault( column.inspector(), {} )[pathString] = inspection
 						inspections.append( inspection )
 
 		if len( inspectors ) == 0 :
@@ -439,12 +437,13 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 
 		tweaks = []
 
-		with Gaffer.Context( self.context() ) as context :
-			for columnSelection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
-				if not isinstance( column, _GafferSceneUI._LightEditorInspectorColumn ) :
-					continue
-				for path in columnSelection.paths() :
-					context["scene:path"] = GafferScene.ScenePlug.stringToPath( path )
+		path = pathListing.getPath().copy()
+		for columnSelection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
+			if not isinstance( column, GafferSceneUI.Private.InspectorColumn ) :
+				continue
+			for pathString in columnSelection.paths() :
+				path.setFromString( pathString )
+				with path.inspectionContext() :
 					inspection = column.inspector().inspect()
 					if inspection is not None and inspection.editable() :
 						source = inspection.source()
@@ -459,7 +458,7 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 							) and
 							( editScope is None or editScope.node().isAncestorOf( source ) )
 						) :
-							tweaks.append( ( path, column.inspector() ) )
+							tweaks.append( ( pathString, column.inspector() ) )
 						else :
 							return []
 					else :
@@ -471,31 +470,33 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 
 		edits = self.__disablableInspectionTweaks( pathListing )
 
-		with Gaffer.UndoScope( self.scriptNode() ), Gaffer.Context( self.context() ) as context :
-			for path, inspector in edits :
-				context["scene:path"] = GafferScene.ScenePlug.stringToPath( path )
+		path = pathListing.getPath().copy()
+		with Gaffer.UndoScope( self.scriptNode() ) :
+			for pathString, inspector in edits :
+				path.setFromString( pathString )
+				with path.inspectionContext() :
+					inspection = inspector.inspect()
+					if inspection is not None and inspection.editable() :
+						source = inspection.source()
 
-				inspection = inspector.inspect()
-				if inspection is not None and inspection.editable() :
-					source = inspection.source()
-
-					if isinstance( source, ( Gaffer.TweakPlug, Gaffer.NameValuePlug ) ) :
-						source["enabled"].setValue( False )
-					elif isinstance( inspector, GafferSceneUI.Private.SetMembershipInspector ) :
-						inspector.editSetMembership( inspection, path, GafferScene.EditScopeAlgo.SetMembership.Unchanged )
+						if isinstance( source, ( Gaffer.TweakPlug, Gaffer.NameValuePlug ) ) :
+							source["enabled"].setValue( False )
+						elif isinstance( inspector, GafferSceneUI.Private.SetMembershipInspector ) :
+							inspector.editSetMembership( inspection, pathString, GafferScene.EditScopeAlgo.SetMembership.Unchanged )
 
 	def __removableAttributeInspections( self, pathListing ) :
 
 		inspections = []
 
-		with Gaffer.Context( self.context() ) as context :
-			for columnSelection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
-				if not isinstance( column, _GafferSceneUI._LightEditorInspectorColumn ) :
-					continue
-				elif not columnSelection.isEmpty() and type( column.inspector() ) != GafferSceneUI.Private.AttributeInspector :
-					return []
-				for path in columnSelection.paths() :
-					context["scene:path"] = GafferScene.ScenePlug.stringToPath( path )
+		path = pathListing.getPath().copy()
+		for columnSelection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
+			if not isinstance( column, GafferSceneUI.Private.InspectorColumn ) :
+				continue
+			elif not columnSelection.isEmpty() and type( column.inspector() ) != GafferSceneUI.Private.AttributeInspector :
+				return []
+			for pathString in columnSelection.paths() :
+				path.setFromString( pathString )
+				with path.inspectionContext() :
 					inspection = column.inspector().inspect()
 					if inspection is not None and inspection.editable() :
 						source = inspection.source()
@@ -533,7 +534,7 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 		for columnSelection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
 			if (
 				not columnSelection.isEmpty() and (
-					not isinstance( column, _GafferSceneUI._LightEditorInspectorColumn ) or
+					not isinstance( column, GafferSceneUI.Private.InspectorColumn ) or
 					not (
 						Gaffer.Metadata.value( "attribute:" + column.inspector().name(), "ui:scene:acceptsSetName" ) or
 						Gaffer.Metadata.value( "attribute:" + column.inspector().name(), "ui:scene:acceptsSetNames" ) or
@@ -726,7 +727,7 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 
 		for i in range( 0, len( columns ) ) :
 			column = columns[ i ]
-			if not isinstance( column, _GafferSceneUI._LightEditorInspectorColumn ) :
+			if not isinstance( column, GafferSceneUI.Private.InspectorColumn ) :
 				continue
 
 			for path in selection[i].paths() :